+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## CHANGELOG
 
+### 3.x.x (2021-MM-DD)
+
+**SDK**
+
+- **Feature:** Added `rakuten.miniapp.user.ACCESS_TOKEN` custom permission
+
+---
 ### 3.0.0 (2021-03-19)
 
 **SDK**